+++ src/__tests__/pages/pupils/index.test.tsx
@@ -1,3 +1,5 @@
+import assert from "assert";
+
 import {
   getStaticPaths,
   getStaticProps,
@@ -7,8 +9,12 @@ import * as curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/ind
 import { allLessonReviewSections } from "@/components/PupilComponents/LessonEngineProvider";
 import pupilLessonOverviewFixture from "@/node-lib/curriculum-api/fixtures/pupilLessonOverview.fixture";
 import { quizQuestions } from "@/node-lib/curriculum-api-2023/fixtures/quizElements.fixture";
+import * as getDownloadResourcesExistence from "@/components/SharedComponents/helpers/downloadAndShareHelpers/getDownloadResourcesExistence";
 
 jest.mock("@/utils/handleTranscript");
+jest.mock(
+  "@/components/SharedComponents/helpers/downloadAndShareHelpers/getDownloadResourcesExistence",
+);
 
 describe("pages/pupils/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]/index", () => {
   describe("pickAvailableSectionsForLesson", () => {
@@ -58,6 +64,18 @@ describe("pages/pupils/programmes/[programmeSlug]/units/[unitSlug]/lessons/[less
   });
 
   describe("getStaticProps", () => {
+    let getDownloadResourcesExistenceSpy: jest.SpyInstance;
+
+    beforeEach(() => {
+      getDownloadResourcesExistenceSpy = jest
+        .spyOn(getDownloadResourcesExistence, "default")
+        .mockResolvedValue({ resources: [] });
+    });
+
+    afterEach(() => {
+      getDownloadResourcesExistenceSpy.mockRestore();
+    });
+
     it("Should call API:pupilLessonOverview on 'pupil'", async () => {
       await getStaticProps({
         params: {
@@ -74,7 +92,6 @@ describe("pages/pupils/programmes/[programmeSlug]/units/[unitSlug]/lessons/[less
         unitSlug: "unitSlug",
         lessonSlug: "lessonSlug",
       });
-      expect(curriculumApi2023.default.pupilLessonOverview).toHaveBeenCalled();
     });
     it("Should call both API::pupilLessonOverview on 'teachers-2023'", async () => {
       await getStaticProps({
@@ -92,13 +109,56 @@ describe("pages/pupils/programmes/[programmeSlug]/units/[unitSlug]/lessons/[less
         unitSlug: "unitSlug",
         lessonSlug: "lessonSlug",
       });
-      expect(
-        curriculumApi2023.default.pupilLessonOverview,
-      ).toHaveBeenCalledWith({
-        programmeSlug: "ks123",
-        unitSlug: "unitSlug",
-        lessonSlug: "lessonSlug",
+    });
+
+    it("tests for the presence of a worksheet", async () => {
+      await getStaticProps({
+        params: {
+          programmeSlug: "ks123",
+          unitSlug: "unitSlug",
+          lessonSlug: "lessonSlug",
+        },
       });
+
+      expect(getDownloadResourcesExistence.default).toHaveBeenCalledWith(
+        "lessonSlug",
+        "worksheet-pdf",
+        false,
+      );
+    });
+
+    it("sets `hasWorksheet` to `true` when a worksheet exists", async () => {
+      getDownloadResourcesExistenceSpy.mockResolvedValue({
+        resources: [["worksheet-pdf", { exists: true }]],
+      });
+
+      const res = await getStaticProps({
+        params: {
+          programmeSlug: "ks123",
+          unitSlug: "unitSlug",
+          lessonSlug: "lessonSlug",
+        },
+      });
+
+      assert("props" in res);
+      expect(res.props.hasWorksheet).toBe(true);
+    });
+
+    it("sets `hasWorksheet` to `false` when a worksheet does not exist", async () => {
+      getDownloadResourcesExistenceSpy.mockResolvedValue({
+        resources: [["worksheet-pdf", { exists: false }]],
+      });
+
+      const res = await getStaticProps({
+        params: {
+          programmeSlug: "ks123",
+          unitSlug: "unitSlug",
+          lessonSlug: "lessonSlug",
+        },
+      });
+
+      assert("props" in res);
+      expect(res.props.hasWorksheet).toBe(false);
     });
   });
 });