+++ src/node-lib/curriculum-api-2023/queries/pupilLessonOverview/pupilLessonOverview.query.ts
@@ -1,7 +1,6 @@
 import { pupilLessonOverviewSchema } from "./pupilLessonOverview.schema";
 
 import errorReporter from "@/common-lib/error-reporter";
-import { LEGACY_COHORT } from "@/config/cohort";
 import OakError from "@/errors/OakError";
 import { Sdk } from "@/node-lib/curriculum-api-2023/sdk";
 
@@ -30,11 +29,5 @@ export const pupilLessonOverviewQuery =
       });
     }
 
-    const isLegacyLicense =
-      !lesson.lessonCohort || lesson.lessonCohort === LEGACY_COHORT;
-
-    return pupilLessonOverviewSchema.parse({
-      ...lesson,
-      isLegacyLicense,
-    });
+    return pupilLessonOverviewSchema.parse(lesson);
   };